refactor departure-favorites/index.ts
